remove futures dependency
This gets rid of a dependency that isn't going to be needed when the
runtime is actually moved to Go. It also means that there are no hidden
behaviors besides those in Rust's std, so we aren't accidentally relying
on them.

This code passes Miri, which is "good enough" as a safety doc when its
going to removed very soon.