Remove "runtime" type + block_on finishes
This commit removes the dubiously-useful "runtime" concept, as Go
manages its own runtime and Rust can't influence it very well.

It also adds a channel-based notification system to the Go side of the
system, making it possible for block_on to terminate.

Some types and variables are poorly named, but I plan on continuing to
restructure the runtime in upcoming commits.