Enable no_std
This enables the #![no_std] attribute for the library crate. To do this,
I changed all the imports to alloc/core and replaced Mutex usage with
UnsafeCell. Note that I did *not* verify my usage of UnsafeCell beyond
"it works with cargo run"; this is not sufficient!

There's still a dependency on alloc; I'd like to get rid of it, but it's
very difficult to work with trait objects or long-lived objects without
it.